+++ lib/salah/juristic.rb
@@ -6,4 +6,8 @@ class Salah::Juristic < Struct.new(:id, :name)
   def self.find_by_id(id)
     Salah.juristics.find{|juristic| juristic.id == id}
   end
+
+  def to_int
+    id
+  end
 end